+++ lib/style_capsule/view_component_helper.rb
@@ -10,7 +10,7 @@ module StyleCapsule
   #
   # Usage in ViewComponent layouts:
   #   def call
-  #     helpers.stylesheet_registrymap_tags
+  #     helpers.stylesheet_registry_tags
   #   end
   #
   # Usage in ViewComponent components:
@@ -60,9 +60,5 @@ module StyleCapsule
     def stylesheet_registry_tags(namespace: nil)
       StyleCapsule::StylesheetRegistry.render_head_stylesheets(helpers, namespace: namespace)
     end
-
-    # @deprecated Use {#stylesheet_registry_tags} instead.
-    #   This method name will be removed in a future version.
-    alias_method :stylesheet_registrymap_tags, :stylesheet_registry_tags
   end
 end